requirements.txt
Pillow
numpy
pytest
